requirements.txt
# Stock Analyzer - Dependencies
# Python 3.11+

# Data
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0

# Technical Analysis
ta>=0.11.0

# UI
streamlit>=1.31.0
plotly>=5.18.0

# Alerts
requests>=2.31.0

# Utils
python-dotenv>=1.0.0
pyarrow>=15.0.0  # For parquet caching

# Logging
loguru>=0.7.2
